Move to monorepo structure
- Add lerna
- Upgrade dependencies
- Remove hook from tree-changes. package
- Add tree-changes-hook package
