+++ .eslintrc
@@ -4,7 +4,7 @@
     "plugin:@typescript-eslint/recommended",
     "plugin:import/typescript",
     "plugin:prettier/recommended",
-    "prettier/standard"
+    "prettier"
   ],
   "parser": "@typescript-eslint/parser",
   "env": {
@@ -13,7 +13,14 @@
   },
   "overrides": [
     {
-      "files": ["test/**/*.ts?(x)"],
+      "files": ["scripts/**/*.ts?(x)"],
+      "rules": {
+        "no-console": "off",
+        "@typescript-eslint/no-var-requires": "off"
+      }
+    },
+    {
+      "files": ["packages/**/test/**/*.ts?(x)"],
       "rules": {
         "no-console": "off",
         "@typescript-eslint/ban-ts-comment": "off"
