+++ demo/package.json
@@ -1,6 +1,6 @@
 {
   "name": "tree-changes-demo",
-  "version": "0.6.0",
+  "version": "0.7.0",
   "description": "",
   "keywords": [],
   "main": "src/index.tsx",
@@ -10,7 +10,7 @@
     "react-scripts": "4.0.1",
     "styled-components": "^5.2.1",
     "styled-minimal": "^1.0.0-7",
-    "tree-changes": "^0.6.0-4",
+    "tree-changes-hook": "^0.7.0",
     "web-vitals": "^0.2.4"
   },
   "devDependencies": {
