+++ package.json
@@ -1,81 +1,39 @@
 {
-  "name": "tree-changes",
-  "version": "0.6.1",
-  "description": "Get changes between two versions of data with similar shape",
-  "author": "Gil Barbara <gilbarbara@gmail.com>",
-  "keywords": [
-    "comparison",
-    "tree"
-  ],
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/gilbarbara/tree-changes.git"
-  },
-  "bugs": {
-    "url": "https://github.com/gilbarbara/tree-changes/issues"
-  },
-  "homepage": "https://github.com/gilbarbara/tree-changes#readme",
-  "main": "lib/index.js",
-  "module": "esm/index.js",
-  "files": [
-    "esm",
-    "lib",
-    "src"
-  ],
-  "types": "lib",
-  "sideEffects": false,
-  "optionalDependencies": {
-    "react": "^16.8.0 || ^17.0.0"
-  },
+  "name": "tree-changes-monorepo",
+  "private": true,
   "dependencies": {
-    "fast-deep-equal": "^3.1.3",
-    "is-lite": "^0.8.0"
-  },
-  "devDependencies": {
     "@gilbarbara/tsconfig": "^0.1.0",
-    "@size-limit/preset-small-lib": "^4.9.1",
-    "@testing-library/jest-dom": "^5.11.6",
-    "@testing-library/react": "^11.2.2",
-    "@types/jest": "^26.0.18",
-    "@types/node": "^14.14.11",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@typescript-eslint/eslint-plugin": "^4.9.1",
-    "@typescript-eslint/parser": "^4.9.1",
+    "@size-limit/preset-small-lib": "^4.9.2",
+    "@types/jest": "^26.0.20",
+    "@types/node": "^14.14.31",
+    "@typescript-eslint/eslint-plugin": "^4.16.1",
+    "@typescript-eslint/parser": "^4.16.1",
     "del-cli": "^3.0.1",
-    "eslint": "^7.15.0",
+    "eslint": "^7.21.0",
     "eslint-config-airbnb-base": "^14.2.1",
-    "eslint-config-prettier": "^7.0.0",
+    "eslint-config-prettier": "^8.1.0",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-prettier": "^3.2.0",
-    "husky": "^4.3.5",
+    "eslint-plugin-prettier": "^3.3.1",
+    "fast-deep-equal": "^3.1.3",
+    "husky": "^4.3.8",
     "jest": "^26.6.3",
+    "lerna": "^3.22.1",
     "prettier": "^2.2.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "repo-tools": "^0.2.0",
-    "size-limit": "^4.9.1",
-    "ts-jest": "^26.4.4",
+    "repo-tools": "^0.2.2",
+    "size-limit": "^4.9.2",
+    "ts-jest": "^26.5.2",
     "ts-node": "^9.1.1",
-    "typescript": "^4.1.2"
+    "typescript": "^4.1.5"
   },
   "scripts": {
-    "build": "npm run clean && npm run build:cjs && npm run build:esm",
-    "build:cjs": "tsc",
-    "build:esm": "tsc -m es6 --outDir esm",
-    "watch:cjs": "npm run build:cjs -- -w",
-    "watch:esm": "npm run build:esm -- -w",
-    "lint": "eslint --ext .ts,.tsx src test",
-    "clean": "del lib/* && del esm/*",
-    "test": "jest",
-    "test:coverage": "jest --coverage --bail",
-    "test:watch": "jest --watchAll --verbose",
-    "format": "prettier \"**/*.{js,jsx,json,yml,yaml,css,less,scss,ts,tsx,md,graphql,mdx}\" --write",
-    "validate": "npm run lint && npm run test:coverage && npm run build && npm run size",
-    "size": "size-limit",
-    "prepublishOnly": "npm run validate"
+    "bootstrap": "lerna bootstrap --hoist",
+    "build": "lerna run build",
+    "lint": "lerna run lint",
+    "size": "lerna run size",
+    "test": "lerna run test:coverage",
+    "validate": "npm run build && npm run lint && npm run test && npm run size",
+    "postinstall": "npm run bootstrap"
   },
   "prettier": {
     "trailingComma": "all",
@@ -84,24 +42,6 @@
     "printWidth": 100,
     "proseWrap": "never"
   },
-  "size-limit": [
-    {
-      "path": "./lib/index.js",
-      "limit": "2.5 kB"
-    },
-    {
-      "path": "./esm/index.js",
-      "limit": "2.5 kB"
-    },
-    {
-      "path": "./lib/hook.js",
-      "limit": "6 kB"
-    },
-    {
-      "path": "./esm/hook.js",
-      "limit": "6 kB"
-    }
-  ],
   "husky": {
     "hooks": {
       "pre-commit": "repo-tools check-remote && npm run validate",
