+++ packages/tree-changes-hook/package.json
@@ -1,0 +1,76 @@
+{
+  "name": "tree-changes-hook",
+  "version": "0.6.1",
+  "description": "Get changes between two versions of data with similar shape",
+  "author": "Gil Barbara <gilbarbara@gmail.com>",
+  "keywords": [
+    "comparison",
+    "tree",
+    "react",
+    "hook"
+  ],
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/gilbarbara/tree-changes.git"
+  },
+  "bugs": {
+    "url": "https://github.com/gilbarbara/tree-changes/issues"
+  },
+  "homepage": "https://github.com/gilbarbara/tree-changes/tree-changes-hook#readme",
+  "main": "lib/index.js",
+  "module": "esm/index.js",
+  "files": [
+    "esm",
+    "lib",
+    "src"
+  ],
+  "types": "lib",
+  "sideEffects": false,
+  "peerDependencies": {
+    "react": "^16.8.0 || ^17.0.0"
+  },
+  "dependencies": {
+    "@testing-library/jest-dom": "^5.11.9",
+    "@testing-library/react": "^11.2.5",
+    "@types/react": "^17.0.2",
+    "@types/react-dom": "^17.0.1",
+    "fast-deep-equal": "^3.1.3",
+    "react": "^16.8.0 || ^17.0.0",
+    "react-dom": "^17.0.1",
+    "tree-changes": "^0.6.1"
+  },
+  "scripts": {
+    "build": "npm run clean && npm run build:cjs && npm run build:esm",
+    "build:cjs": "tsc",
+    "build:esm": "tsc -m es6 --outDir esm",
+    "watch:cjs": "npm run build:cjs -- -w",
+    "watch:esm": "npm run build:esm -- -w",
+    "lint": "eslint --ext .ts,.tsx src test",
+    "clean": "del lib/* && del esm/*",
+    "test": "jest",
+    "test:coverage": "jest --coverage --bail",
+    "test:watch": "jest --watchAll --verbose",
+    "format": "prettier \"**/*.{js,jsx,json,yml,yaml,css,less,scss,ts,tsx,md,graphql,mdx}\" --write",
+    "validate": "npm run lint && npm run test:coverage && npm run build && npm run size",
+    "size": "size-limit",
+    "prepublishOnly": "npm run validate"
+  },
+  "prettier": {
+    "trailingComma": "all",
+    "singleQuote": true,
+    "arrowParens": "avoid",
+    "printWidth": 100,
+    "proseWrap": "never"
+  },
+  "size-limit": [
+    {
+      "path": "./lib/index.js",
+      "limit": "3 kB"
+    },
+    {
+      "path": "./esm/index.js",
+      "limit": "3 kB"
+    }
+  ]
+}
