+++ packages/tree-changes/package.json
@@ -1,0 +1,58 @@
+{
+  "name": "tree-changes",
+  "version": "0.6.1",
+  "description": "Get changes between two versions of data with similar shape",
+  "author": "Gil Barbara <gilbarbara@gmail.com>",
+  "keywords": [
+    "comparison",
+    "tree"
+  ],
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/gilbarbara/tree-changes.git"
+  },
+  "bugs": {
+    "url": "https://github.com/gilbarbara/tree-changes/issues"
+  },
+  "homepage": "https://github.com/gilbarbara/tree-changes#readme",
+  "main": "lib/index.js",
+  "module": "esm/index.js",
+  "files": [
+    "esm",
+    "lib",
+    "src"
+  ],
+  "types": "lib",
+  "sideEffects": false,
+  "dependencies": {
+    "fast-deep-equal": "^3.1.3",
+    "is-lite": "^0.8.0"
+  },
+  "scripts": {
+    "build": "npm run clean && npm run build:cjs && npm run build:esm",
+    "build:cjs": "tsc",
+    "build:esm": "tsc -m es6 --outDir esm",
+    "watch:cjs": "npm run build:cjs -- -w",
+    "watch:esm": "npm run build:esm -- -w",
+    "lint": "eslint --ext .ts,.tsx src test",
+    "clean": "del lib/* && del esm/*",
+    "test": "jest",
+    "test:coverage": "jest --coverage --bail",
+    "test:watch": "jest --watchAll --verbose",
+    "format": "prettier \"**/*.{js,jsx,json,yml,yaml,css,less,scss,ts,tsx,md,graphql,mdx}\" --write",
+    "validate": "npm run lint && npm run test:coverage && npm run build && npm run size",
+    "size": "size-limit",
+    "prepublishOnly": "npm run validate"
+  },
+  "size-limit": [
+    {
+      "path": "./lib/index.js",
+      "limit": "3 kB"
+    },
+    {
+      "path": "./esm/index.js",
+      "limit": "3 kB"
+    }
+  ]
+}
