+++ packages/tree-changes/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "@gilbarbara/tsconfig",
+  "compilerOptions": {
+    "outDir": "./lib",
+    "target": "es5"
+  },
+  "include": ["src/**/*"],
+  "exclude": ["node_modules", "**/*.spec.ts"]
+}
